删除原steps组件，将新版step使用steps重命名
