--- src/app/demo/pc/step/basic/demo.component.html
@@ -1,9 +1,0 @@
-<!-- ignore the following lines, they are not important to this demo -->
-<jigsaw-demo-description [summary]="summary" [content]="description">
-</jigsaw-demo-description>
-
-<jigsaw-header [level]="2">这是step的最基础用法</jigsaw-header>
-<jigsaw-step [data]="data" [current]="current"></jigsaw-step>
-<div>
-    <jigsaw-button (click)="changeData()">换一批数据瞅瞅</jigsaw-button>
-</div>
