--- src/app/demo/pc/step/basic/demo.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from "@angular/core";
-import { CommonModule } from "@angular/common";
-import { JigsawDemoDescriptionModule } from "app/demo-description/demo-description";
-import { JigsawStepBasicDemoComponent } from "./demo.component";
-import {
-    JigsawStepModule,
-    JigsawHeaderModule,
-    JigsawButtonModule
-} from "jigsaw/public_api";
-
-@NgModule({
-    imports: [
-        JigsawStepModule,
-        CommonModule,
-        JigsawDemoDescriptionModule,
-        JigsawHeaderModule,
-        JigsawButtonModule
-    ],
-    declarations: [JigsawStepBasicDemoComponent],
-    exports: [JigsawStepBasicDemoComponent]
-})
-export class JigsawStepBasicDemoModule {}
