--- src/app/demo/pc/step/over-length/demo.component.scss
@@ -1,10 +1,0 @@
-#step1 {
-    width: 80%;
-    margin: 10px auto;
-}
-
-#step2 {
-    width: 200px;
-    height: 500px;
-    margin: 10px auto;
-}
