--- src/app/demo/pc/step/vertical/demo.component.html
@@ -1,7 +1,0 @@
-<!-- ignore the following lines, they are not important to this demo -->
-<jigsaw-demo-description [summary]="summary" [content]="description">
-</jigsaw-demo-description>
-
-<jigsaw-header [level]="2">这是step的纵向用法 </jigsaw-header>
-<jigsaw-step [data]="data" [current]="current" direction="vertical"></jigsaw-step>
-<jigsaw-button (click)="changeData()">换一批数据瞅瞅</jigsaw-button>
