+++ src/app/demo/pc/steps/basic/demo.component.scss
@@ -1,4 +1,4 @@
-jigsaw-step {
+jigsaw-steps {
     width: 80%;
     margin: 10px auto;
 }
