--- src/app/demo/pc/steps/custom-icons/demo.component.html
@@ -1,17 +1,0 @@
-<!-- ignore the following lines, they are not important to this demo -->
-<jigsaw-demo-description [summary]="summary" [content]="description">
-</jigsaw-demo-description>
-
-
-<!-- start to learn the demo from here -->
-<jigsaw-steps preSize="large">
-    <jigsaw-step-item *ngFor="let step of steps" [status]="step.status"
-                      [waitingIcon]="step.waitingIcon"
-                      [doneIcon]="step.doneIcon"
-                      [processingIcon]="step.processingIcon"
-                      [errorIcon]="step.errorIcon"
-                      [warningIcon]="step.warningIcon"
-                      [skippedIcon]="step.skippedIcon">
-        {{step.title}}
-    </jigsaw-step-item>
-</jigsaw-steps>
