+++ src/app/demo/pc/steps/over-length/demo.component.html
@@ -3,6 +3,6 @@
 </jigsaw-demo-description>
 
 <jigsaw-header [level]="2">当横向步骤的标题和描述过长时的表现</jigsaw-header>
-<jigsaw-step id="step1" [data]="data" [current]="current"></jigsaw-step>
+<jigsaw-steps width="80%" [data]="data" [current]="current"></jigsaw-steps>
 <jigsaw-header [level]="2">当纵向步骤的标题和描述过长时的表现</jigsaw-header>
-<jigsaw-step id="step2" [data]="data" [current]="current" direction="vertical"></jigsaw-step>
+<jigsaw-steps width="200" height="350" [data]="data" [current]="current" direction="vertical"></jigsaw-steps>
