+++ src/app/demo/pc/steps/over-length/demo.component.scss
@@ -1,0 +1,3 @@
+jigsaw-steps {
+    margin: 10px auto;
+}
