--- src/app/demo/pc/steps/steps-multiline/demo.component.html
@@ -1,17 +1,0 @@
-<!-- ignore the following lines, they are not important to this demo -->
-<jigsaw-demo-description [summary]="summary" [content]="description">
-</jigsaw-demo-description>
-
-
-<!-- start to learn the demo from here -->
-<j-button (click)="_$handleAdd()"> 每行放更多个</j-button>
-<j-button (click)="_$handleSub()" [disabled]="numInLine == 1 "> 每行放更少个</j-button>
-<j-button (click)="_$handleLarge()"> 大图标</j-button>
-<j-button (click)="_$handleSmall()"> 小图标</j-button>
-<j-button (click)="_$handleDefault()"> 默认图标</j-button>
-<div> 一行有 {{numInLine}} 个元素</div>
-<div style="height: calc( 100vh - 110px )">
-    <jigsaw-steps-multiline [data]="steps" [numInline]="numInLine" [preSize]="presize"
-                            (select)="_$selectChange($event)">
-    </jigsaw-steps-multiline>
-</div>
