+++ src/app/demo/pc/steps/vertical/demo.component.html
@@ -2,11 +2,6 @@
 <jigsaw-demo-description [summary]="summary" [content]="description">
 </jigsaw-demo-description>
 
-
-<!-- start to learn the demo from here -->
-<jigsaw-steps  direction="vertical"  width="200px">
-    <jigsaw-step-item *ngFor="let step of steps" [status]="step.status">
-        <div jigsaw-title>{{step.title}}</div>
-        <div jigsaw-sub-v-title>{{step.subTitle}}</div>
-    </jigsaw-step-item>
-</jigsaw-steps>
+<jigsaw-header [level]="2">这是step的纵向用法 </jigsaw-header>
+<jigsaw-steps [data]="data" [current]="current" direction="vertical"></jigsaw-steps>
+<jigsaw-button (click)="changeData()">换一批数据瞅瞅</jigsaw-button>
