+++ src/app/demo/pc/steps/vertical/demo.component.scss
@@ -1,4 +1,4 @@
-jigsaw-step {
+jigsaw-steps {
     width: 200px;
     height: 400px;
     margin: 10px;
