--- src/jigsaw/pc-components/step/step.html
@@ -1,37 +1,0 @@
-<div
-    *ngFor="let item of data;index as index"
-    class="jigsaw-step-item"
-    [ngClass]="{'jigsaw-step-item-waiting': current < index,
-                'jigsaw-step-item-progressing': current === index,
-                'jigsaw-step-item-finished': current > index,
-                'jigsaw-step-item-normal': item?.status === 'normal',
-                'jigsaw-step-item-error': item?.status === 'error',
-                'jigsaw-step-item-warning': item?.status === 'warning'
-                }"
->
-    <div class="jigsaw-step-title-cntr">
-        <span class="jigsaw-step-icon">
-            <i
-                class="iconfont"
-                *ngIf="current > index"
-                [ngClass]="{'iconfont-ea37':item?.status === 'normal',
-                        'iconfont-e192':item?.status === 'error',
-                        'iconfont-e76c':item?.status === 'warning'
-                    }"
-            ></i>
-            <span class="jigsaw-step-icon-number" *ngIf="current <= index"
-                >{{index}}</span
-            >
-        </span>
-        <span class="jigsaw-step-title" title="{{item?.title}}">{{item?.title}}</span>
-        <span class="jigsaw-step-bar" *ngIf="direction === 'horizontal'"></span>
-    </div>
-    <div
-        class="jigsaw-step-sub-title"
-        trustedHtml="{{item?.subTitle}}" 
-        [trustedHtmlContext]="item?.context"
-    ></div>
-    <div class="jigsaw-step-bar" *ngIf="direction === 'vertical'">
-        <span></span>
-    </div>
-</div>
