+++ src/jigsaw/pc-components/theming/all-theme.scss
@@ -23,9 +23,7 @@
 @import "../radio/radio";
 @import "../select/select";
 @import "../slider/slider";
-@import "../step/step";
-@import "../steps/steps";
-@import "../steps/steps-multiline";
+@import "../steps/step";
 @import "../switch/switch";
 @import "../table/table";
 @import "../table/table-head";
