+++ src/jigsaw/public_api.ts
@@ -67,9 +67,7 @@ export * from "./pc-components/rate/index";
 export * from "./pc-components/root/root";
 export * from "./pc-components/select/select";
 export * from "./pc-components/slider/index";
-export * from "./pc-components/step/step";
-export * from "./pc-components/steps/index";
-export * from "./pc-components/steps/steps-multiline";
+export * from "./pc-components/steps/step";
 export * from "./pc-components/switch/index";
 export * from "./pc-components/table/table-inner.components";
 export * from "./pc-components/table/table-renderer";
